Add GPX as a serial port filter
The idea is to have GPX translate between g-codes and x3g.  GPX
already does a pretty good job for output gcode intended for print
instructions, that is the gcode output by slicers.
It has some minimal support for talking directly to the printer and
receiving data back from the printer.  What I'm hoping to add is
translation of printer output to responses a generic Marlin printer
might produce.  For now, I'm focused on the I/O that OctoPrint uses.