Initial SDCardFileStorage implementation of StorageInterface
This initial change is intended to create a StorageInterface
implementation for the sdcard. At this point it doesn't attempt to
change any functionality outside of that already supported for SD cards,
just change the code flow through the fileManager.

Is this the direction you were talking about in issue #851?  What way
would you prefer to handle the unsupported (yet) functionality.
NotImplemented as here or with io.Unsupported as specified for some?

Also, note how I've got this calling back through the fileManager for
LOCAL storage on upload. Should it work this way or should it be a
subclass of LocalFileStorage or some other structure? To do some of the
other StorageInterface, more of it will want to rely on the
LocalFileStorage, I think.

Perhaps the SDCardFileStorage should create its own LocalFileStorage at
a different base path as an SDCard shadow to do some of the metadata,
etc. As an actual mirror rather than just for upload, it would be more
reliable for some of the metadata, analysis purposes. It could also be
enhanced if firmwares were willing to give us a hash for comparison (to
avoid streaming the file to check).