Add S255 parameter to M106 for "Fan On"
Bit of a debate with some gcode purists. RepRapPro demands an S
parameter.  Adding S255 should be backward compatible with the firmwares
that recognize M106 as full on. As with other issues, it'd be nice to
have something more to go on here than the reprap wiki and firmware
source code.