Auto uppercase Gcode except commands in configurable blacklist
Added new config setting: auto uppercase blacklist that is configurable in the advanced option of the serial tab, with M117 included in the default settings of the blacklist.
Changed the sendCommand method in the TerminalViewModel to auto uppercase the entire gcode command except when gcode is in the blacklist.
Updated documentaion to include the defualt auto uppercase blacklist command M117.

Resolves: #1026