First throw at securing the socket & propagating user changes
The socket client will now authenticate with the socket after login,
using the session id from the login response. This will associate the
session with the socket server side. Any messages to the client apart
from "connected" and "reauthRequired", will be checked against the
STATUS permission and not be sent if that is missing from the current
user (which might also be the anonymous/guest user).

A logout of the session and any changes of the user or their permissions
will trigger trigger a reauthRequired message to be sent over the
socket, instructing the client to either perform a new passive login to
update the user's permissions, or to logout altogether (server side
logout or user removal).

Known bugs:

  * Changing group permissions sometimes doesn't propagate through -
    seems to be caused by the code not properly finding the login
    sessions of affected users based on the changed group. Not yet sure
    what's up there.

Open points:

  * Deactivation of a user's account also needs to invalidate the
    session
  * Session timeout needs to be evaluated (currently 30min, unclear if
    that suffices)
  * possibly more I can't think of right now