Remove branch from display version
It's no longer reliable since the Github archive export we'll now
usually use for upgrading does weird things when a tag is available on
more than one branch and we don't want our current stable version to
be reported as "rc/devel" just because that tag is available there too.

A full blown version number should be definitive enough considering that
if it's not a tag it will also contain the commit hash.