Don't write default update type & method to config
If we set method to pip in 1.3.6 and then roll back to 1.3.5 we will
need to a manual config.yaml update otherwise or we will no longer be
able to update.

Be removing the force parameter and using the actual default settings
instead of current ones, we don't do that anymore.