Revert "Update to pyserial 3.4"
This is causing issues on update thanks to the PySerial installation
logging an error if the former version was installed through
dependency resolution during OctoPrint's python setup.py install run
(which is most likely the case for pretty much all live installations
out there).

Update (and hence activation of fix for #2333) postponed until a
solution for this issue is found.

This reverts commit 77989d2a763ac94ef0367a5a735db5ee8bc6ca40.