SWU: Workaround for pip problem with eggs
Starting with pip 7.x pip runs into an issue if a software package
formerly installed as a Python egg (= zip file stored in site-packages,
referenced through easy-install.pth, e.g. through a stand alone "python
setup.py install") is to be updated to a newer version.

The issue arises since the egg is still on the python path after the
replacement has taken place, and pip tries to read from it without error
handling for the fact that this file is now not longer existing. The
source of the problem appears to be with both the bundled as well as the
stock versions of pkg_resources and pip's usage of them to perform its
management duties. It's currently unclear if the problem is a bug or a
feature, even though it strongly looks like the first one.

In any case, during tests it was found that this problem can be worked
around by simply performing a second "pip install" with same parameters
right after the failed first one.

This patch detects this scenario based on the logged error message and
performs the second install if deemed necessary.

This is a preparational step to allow upgrade of the pyserial dependency
(and other dependencies currently installed as eggs as part of
OctoPrint's former "python setup.py install" installation procedure)
with OctoPrint 1.3.8, to finalize the fix for #2333.