comm: custom readline implementation w/ buffer
We always read everything from the serial that is currently available in
our rx buffer but only return the first line in that. While that
increases our memory consumption slightly, it should hopefully not cause
issues and should greatly reduce the risk of buffer overflows making us
lose or receive corrupt data.

Possible solution for #2370 and #2454