add plugin exception tracking
one of the things that we're seeing is that outdated plugins are breaking Octoprint instances. Given an RC is installed by ~1000 people and about ~20-30 speak up about running it, it's a pretty safe bet that there are fairly large-scale plugin failures that are barely surfacing on github issues or discourse.

this doesn't catch ALL exceptions, just ones that come from plugins.

tested  by adding a `raise` to the cura plugin at load time. kinda fun:

    2018-12-23 19:59:46,312 - octoprint.plugins.tracking - INFO - Sent tracking event server_exception, payload: {'path': u'plugin/cura/on_startup', 'exception_trace': 'Traceback (most recent call last):\n  File "/home/user/devel/OctoPrint/src/octoprint/plugin/__init__.py", line 233, in call_plugin\n    result = getattr(plugin, method)(*args, **kwargs)\n  File "/home/user/devel/OctoPrint/src/octoprint/plugins/cura/__init__.py", line 86, in on_startup\n    raise Exception("ted quick fail test")\nException: ted quick fail test\n', 'plugin_name': u''}

- carefully look for event object and raise an event if a problem is caught
- add documentation for why 'events' is injected