Fixup actual temperature parsing for CR-10 S Pro
The temperature updates from this printer look like this:

```
Recv:  ==T:20.00 /0.00 ==B:21.02 /0.00 @:0 B@:0
```

The `regex_temp` knows how to extract the data from this line, but there is some different logic that gates running the regex that needs to know what else to look for.

In the longer term, should this line really just execute `regex_temp`?  Regardless, for now, the most expedient change to enable the temperature graph is to expand the logic as in this diff.